Moved coins and loading state hooks from CryptoTable to CryptoContext.
