+++ src/components/CoinsTable.js
@@ -16,8 +16,6 @@ import {
   Table,
   Paper,
 } from "@material-ui/core";
-import axios from "axios";
-import { CoinList } from "../config/api";
 import { useNavigate } from "react-router-dom";
 import { CryptoState } from "../CryptoContext";
 
@@ -26,12 +24,10 @@ export function numberWithCommas(x) {
 }
 
 export default function CoinsTable() {
-  const [loading, setLoading] = useState(false);
-  const [coins, setCoins] = useState([]);
   const [search, setSearch] = useState("");
   const [page, setPage] = useState(1);
 
-  const { currency, symbol } = CryptoState();
+  const { currency, symbol, coins, loading, fetchCoins } = CryptoState();
 
   const useStyles = makeStyles({
     row: {
@@ -61,14 +57,14 @@ export default function CoinsTable() {
     },
   });
 
-  const fetchCoins = async () => {
-    setLoading(true);
-    const { data } = await axios.get(CoinList(currency));
-    console.log(data);
+  // const fetchCoins = async () => {
+  //   setLoading(true);
+  //   const { data } = await axios.get(CoinList(currency));
+  //   console.log(data);
 
-    setCoins(data);
-    setLoading(false);
-  };
+  //   setCoins(data);
+  //   setLoading(false);
+  // };
 
   useEffect(() => {
     fetchCoins();
